Move SessionPoolEntry into SessionPoolImpl and make it private
(That's the only use of it)
